+++ NEWS.md
@@ -1,6 +1,7 @@
 # argonR 0.2.0.9000
+- Fix #9: remove close button from `argonModal()`. Thanks @Morais14
 - change the icon parameter of `argonTabSet()`, `argonCard()`, `argonAlert()`,
-`argonDropdownItem()` and `argonNavItem()`
+`argonDropdownItem()`, `argonNavItem()` and `argonButton()`
 Now you can use `icon("cogs")` from shiny or `argonIcon()` as usual
 - add offset arg to `argonColumn()`
 - Add btn_text argument to `argonCard()`